requirements.txt
gradio==<latest_version>
pdfplumber==<latest_version>
google-generativeai==<latest_version>
requests==<latest_version>
